backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
bm25s
snowballstemmer
aiosqlite
python-multipart
langdetect
numpy
python-jose[cryptography]
passlib[bcrypt]
bcrypt
